update admin data

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
 // use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
@@ -94,7 +95,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+
+        if(Gate::denies('update-post', $post))
+        {
+            abort(403, 'You cannot edit this blog post!');
+        }
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -107,6 +115,14 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if(Gate::denies('update-post', $post))
+        // {
+        //     abort(403, 'You cannot edit this blog post!');
+        // }
+
+        $this->authorize('update-post', $post);
+
         $validated = $request->validated(); 
         $post->fill($validated);
         $post->save();
@@ -125,6 +141,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if(Gate::denies('delete-post', $post))
+        // {
+        //     abort(403, 'You cannot delete this blog post!');
+        // }
+
+        $this->authorize('delete-post', $post);
+
         $post->delete();
 
         session()->flash('status', 'blog post was delete!');
